fix(fastify): remove duplicate done variable in multipart parser
