+++ server-adapters/fastify/src/multipart-limit.test.ts
@@ -1,0 +1,62 @@
+import Fastify from 'fastify';
+import { describe, it, expect } from 'vitest';
+import { MastraServer } from './index';
+
+describe('Multipart Size Limit (Bug Fix)', () => {
+  it('should reject when file exceeds maxFileSize', async () => {
+    const app = Fastify();
+
+    const adapter = new MastraServer({
+      app,
+      mastra: {
+        getLogger: () => ({
+          error: () => {}, // stub logger to avoid crashes
+        }),
+      } as any,
+      bodyLimitOptions: { maxSize: 100 }, // very small limit
+    });
+
+    // Register middleware
+    app.addHook('preHandler', adapter.createContextMiddleware());
+
+    // Register route
+    await adapter.registerRoute(
+      app,
+      {
+        method: 'POST',
+        path: '/upload',
+        responseType: 'json',
+        handler: async () => ({ ok: true }),
+      },
+      { prefix: '' },
+    );
+
+    const address = await app.listen({ port: 0 });
+
+    try {
+      // Create multipart payload exceeding limit
+      const boundary = '----testboundary';
+      const bigData = 'a'.repeat(1000); // exceeds 100 bytes
+
+      const payload =
+        `--${boundary}\r\n` +
+        `Content-Disposition: form-data; name="file"; filename="test.txt"\r\n` +
+        `Content-Type: text/plain\r\n\r\n` +
+        `${bigData}\r\n` +
+        `--${boundary}--`;
+
+      const res = await fetch(`${address}/upload`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': `multipart/form-data; boundary=${boundary}`,
+        },
+        body: payload,
+      });
+
+      // Should fail due to file size limit
+      expect(res.status).toBeGreaterThanOrEqual(400);
+    } finally {
+      await app.close();
+    }
+  });
+});
